// src/lib/api/generated/models/user.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */
import type { UserStatus } from './userStatus';

export interface User {
  created_at: string;
  email: string;
  /** @nullable */
  email_verified_at?: string | null;
  id: number;
  last_login?: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  profile_image?: string | null;
  registration_date?: string;
  status: UserStatus;
  /** @nullable */
  suspension_reason?: string | null;
  /** @nullable */
  suspension_until?: string | null;
  total_auctions: number;
  total_bids: number;
  total_winnings: number;
  updated_at: string;
}