// src/lib/api/generated/models/updateUserRequest.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */
import type { UpdateUserRequestStatus } from './updateUserRequestStatus';

export interface UpdateUserRequest {
  email?: string;
  name?: string;
  phone?: string;
  status?: UpdateUserRequestStatus;
}