// src/lib/api/generated/models/getAdminUsersParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */
import type { GetAdminUsersStatus } from './getAdminUsersStatus';
import type { GetAdminUsersSortDirection } from './getAdminUsersSortDirection';

export type GetAdminUsersParams = {
page?: number;
per_page?: number;
search?: string;
status?: GetAdminUsersStatus;
sort_by?: string;
sort_direction?: GetAdminUsersSortDirection;
};